# ips_autobot_svr/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ips_autobot_svr API",
    "version": "1.0.0",
    "description": "Admin and user auth, logs, recharge APIs. Multi-app zhiling/yifei — pass app where required. Password fields use RSA encryption (GET /api/public-key)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "List users (admin only)",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "username",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "phone",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ]
            }
          },
          {
            "in": "query",
            "name": "created_at_from",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "created_at_to",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "last_login_from",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "last_login_to",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "username": {
                                "type": "string",
                                "nullable": true
                              },
                              "phone": {
                                "type": "string"
                              },
                              "status": {
                                "type": "integer"
                              },
                              "created_at": {
                                "type": "string"
                              },
                              "last_login_at": {
                                "type": "string",
                                "nullable": true
                              },
                              "clients": {
                                "type": "object",
                                "description": "Per-app zhiling and yifei slices (MAC, score, tiers, member_type, member_expire_at)",
                                "properties": {
                                  "zhiling": {
                                    "type": "object"
                                  },
                                  "yifei": {
                                    "type": "object"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create user",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "phone",
                  "passwordEncrypted"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "User display name"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "passwordEncrypted": {
                    "type": "string"
                  },
                  "member_type": {
                    "type": "integer",
                    "enum": [
                      0,
                      1
                    ],
                    "description": "Initial member_type for zhiling client only; yifei starts as 0"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created; response includes clients.zhiling and clients.yifei defaults"
          },
          "409": {
            "description": "Phone already registered"
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "summary": "User login (by phone or username)",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "passwordEncrypted"
                ],
                "properties": {
                  "phone": {
                    "type": "string",
                    "description": "Phone number (use either phone or username)"
                  },
                  "username": {
                    "type": "string",
                    "description": "Username (use either phone or username)"
                  },
                  "passwordEncrypted": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login success; user object has shared fields and clients.zhiling / clients.yifei",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "username": {
                              "type": "string",
                              "nullable": true
                            },
                            "phone": {
                              "type": "string"
                            },
                            "status": {
                              "type": "integer"
                            },
                            "created_at": {
                              "type": "string"
                            },
                            "clients": {
                              "type": "object"
                            },
                            "last_login_at": {
                              "type": "string",
                              "format": "date-time"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Phone or username and password required"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "summary": "Get current user profile",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user profile"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/password": {
      "put": {
        "summary": "Change password (no old password required; user self or admin for another user)",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "newPasswordEncrypted"
                ],
                "properties": {
                  "newPasswordEncrypted": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "integer",
                    "description": "Admin only: target user id"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated"
          }
        }
      }
    },
    "/api/users/mac": {
      "get": {
        "summary": "Get bound MAC for current user or admin-specified user",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "app",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "zhiling",
                "yifei"
              ]
            }
          },
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "integer"
            },
            "description": "Required when caller is admin"
          }
        ],
        "responses": {
          "200": {
            "description": "Current mac_addr for app or null"
          },
          "400": {
            "description": "Missing userId for admin"
          }
        }
      },
      "put": {
        "summary": "Bind or update MAC (globally unique)",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "mac_addr",
                  "app"
                ],
                "properties": {
                  "app": {
                    "type": "string",
                    "enum": [
                      "zhiling",
                      "yifei"
                    ]
                  },
                  "mac_addr": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "integer",
                    "description": "Admin only — target user id"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MAC bound for app"
          },
          "400": {
            "description": "Invalid MAC or missing userId for admin"
          },
          "409": {
            "description": "MAC already used by another user"
          }
        }
      }
    },
    "/api/users/mac/verify": {
      "post": {
        "summary": "Verify MAC matches bound value (user JWT only)",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "mac_addr",
                  "app"
                ],
                "properties": {
                  "app": {
                    "type": "string",
                    "enum": [
                      "zhiling",
                      "yifei"
                    ]
                  },
                  "mac_addr": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "matched true/false per app; unbound returns matched false"
          }
        }
      }
    },
    "/api/users/{id}/mac": {
      "delete": {
        "summary": "Unbind user MAC (super admin only)",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Target user id"
          },
          {
            "in": "query",
            "name": "app",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "zhiling",
                "yifei"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MAC cleared for app",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "mac_addr": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Super admin required"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
    "/api/users/{id}/disable": {
      "put": {
        "summary": "Disable user (admin only)",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User disabled"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
...
}